backend/app/requirements.txt
# /app/requirements.txt

fastapi
uvicorn
python-multipart

# Base de Datos y ORM
sqlalchemy
psycopg2-binary

# Configuración y Validación
# pydantic[email] instalará pydantic junto con la dependencia 'email-validator'
pydantic[email]
pydantic-settings

# Seguridad y Autenticación
python-jose[cryptography]
passlib
bcrypt

# Documentación
mkdocs
mkdocs-material
PyYAML

# Utilidades (Git)
GitPython
aiofiles
